requirements.txt
Django
djangorestframework
django-cors-headers
psycopg2-binary
python-decouple
django-extensions
Pillow
whitenoise